# openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele API
  version: "1.0.0"
  description: Wedding gift management system with family tree, escrow-based pool funding, and currency snapshot locking.
servers:
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────────────────
  # Standard responses (переиспользуемые)
  # ─────────────────────────────────────
  responses:
    BadRequest:
      description: Bad request — invalid input or business rule violation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request body validation failed (Joi)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─────────────────────────────
    # Weddings
    # ─────────────────────────────
...
    CreateContributionRequest:
      type: object
      required: [poolId, originalAmount, originalCurrency, idempotencyKey]
      properties:
        poolId:
          type: integer
          example: 1
        originalAmount:
          type: number
          example: 100
        originalCurrency:
          type: string
          example: "USD"
        idempotencyKey:
          type: string
          example: "uniq-key-12345"
    CreateContributionResponse:
      type: object
      properties:
        id:
          type: integer
        guestId:
          type: integer
        poolId:
          type: integer
        amountKzt:
          type: integer
        originalAmount:
          type: number
        originalCurrency:
          type: string
        exchangeRate:
          type: number
        lockedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [COMPLETED, PENDING, FAILED, REFUNDED]

        ContributionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        guestId:
          type: integer
          example: 3
        poolId:
          type: integer
          example: 1
        amountKzt:
          type: integer
          example: 47050
        originalAmount:
          type: number
          example: 100
        originalCurrency:
          type: string
          example: "USD"
        exchangeRate:
          type: number
          example: 470.5
        lockedAt:
          type: string
          format: date-time
          example: "2025-06-15T14:22:00.000Z"
        status:
          type: string
          example: "COMPLETED"
        createdAt:
          type: string
          format: date-time
          example: "2025-06-15T14:22:00.000Z"
        guest:
          type: object
          nullable: true
          properties:
            id:
              type: integer
              example: 3
            fullName:
              type: string
              example: "Марат Асқарұлы"
    PaginatedContributions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContributionResponse"
        pagination:
          $ref: "#/components/schemas/CursorPagination"

    # ─────────────────────────────
    # Family Tree
    # ─────────────────────────────
...
    UpdateExchangeRateRequest:
      type: object
      required: [currencyFrom, currencyTo, rate]
      properties:
        currencyFrom:
          type: string
          example: "USD"
        currencyTo:
          type: string
          example: "KZT"
        rate:
          type: number
          example: 480.5
        source:
          type: string
          default: "manual"

        AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
          example: 42
        userId:
          type: integer
          example: 1
        action:
          type: string
          example: "UPDATE_EXCHANGE_RATE"
        entityType:
          type: string
          example: "ExchangeRate"
        entityId:
          type: integer
          example: 5
        oldValue:
          type: object
          example: { "rate": 450.0 }
        newValue:
          type: object
          example: { "rate": 470.5 }
        ipAddress:
          type: string
          example: "192.168.1.100"
        createdAt:
          type: string
          format: date-time
          example: "2025-06-10T09:15:00.000Z"
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
            fullName:
              type: string
              example: "Super Admin"
            email:
              type: string
              example: "admin@saukele.kz"
    PaginatedAuditLog:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        pagination:
          $ref: "#/components/schemas/CursorPagination"
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        pagination:
          $ref: "#/components/schemas/PagePagination"

    # ─────────────────────────────
    # Pagination
    # ─────────────────────────────
    CursorPagination:
      type: object
      properties:
        limit:
          type: integer
          example: 10
        nextCursor:
          type: integer
          nullable: true
          example: 42
        hasNextPage:
          type: boolean
          example: true
    PagePagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3

    # ─────────────────────────────
    # Common
    # ─────────────────────────────
        ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Email already registered"
        statusCode:
          type: integer
          description: HTTP status code
          example: 409
        details:
          type: array
          items:
            type: string
          description: Optional validation error details
          example: ["Email 'test@test.com' is already in use"]

        MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Email verified successfully"
paths:
...
  /contributions:
    post:
      tags: [Contributions]
      summary: Make a contribution to a gift pool (locks exchange rate)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContributionRequest"
      responses:
        "201":
          description: Contribution created with locked exchange rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateContributionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /contributions/my:
    get:
      tags: [Contributions]
      summary: Get current user's contributions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: cursor
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of user's contributions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedContributions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /contributions/pool/{poolId}:
    get:
      tags: [Contributions]
      summary: List contributions for a pool (cursor-based pagination)
      security:
        - bearerAuth: []
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: cursor
          in: query
          schema:
            type: integer
          description: Cursor from previous response (nextCursor)
      responses:
        "200":
          description: Paginated contributions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedContributions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /contributions/{id}:
    delete:
      tags: [Contributions]
      summary: Delete a contribution (SUPER_ADMIN only)
      description: Refunds the contribution amount to the pool.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description:
            /contributions/{id}:
    delete:
      tags: [Contributions]
      summary: Delete a contribution (SUPER_ADMIN only)
      description: Refunds the contribution amount to the pool.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contribution cancelled and refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────
  # Family Tree
  # ─────────────────────────────────────
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users with optional filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: role
          schema:
            type: string
            enum: [GUEST, FAMILY_MEMBER, COUPLE, MODERATOR, SUPER_ADMIN]
        - in: query
          name: isBlocked
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/users/{id}:
    delete:
      tags: [Admin]
      summary: Delete a user (SUPER_ADMIN only)
      description: Anonymizes their contributions and removes them from family trees. Cannot delete yourself.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/exchange-rates:
    put:
      tags: [Admin]
      summary: Set a new exchange rate
      description: Expires the current active rate and inserts a new one. Logged in audit trail.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExchangeRateRequest"
      responses:
        "201":
          description: Exchange rate created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/audit-log:
    get:
      tags: [Admin]
      summary: View full audit log with filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: cursor
          schema:
            type: integer
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: entityType
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Paginated audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAuditLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/moderators/{userId}/promote:
    patch:
      tags: [Admin]
      summary: Toggle MODERATOR role for a user
      description: If user is currently MODERATOR, demotes to COUPLE. If not, promotes to MODERATOR.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────
  # Moderator
  # ─────────────────────────────────────
...